fix: حل 4 أخطاء Kotlin compilation
- HomeScreen: إضافة @Composable على clickableNoRipple + import clickable/MutableInteractionSource
- XrayController: إزالة .toLong() من startLoop (Int وليس Long)
- XrayVpnService: تغيير setup() إلى startup() حسب CoreCallbackHandler

+++ app/src/main/kotlin/com/vpnpro/vpn/XrayController.kt
@@ -31,7 +31,7 @@ object XrayController {
         stopInternal()
         try {
             val ctrl = Libv2ray.newCoreController(handler)
-            ctrl.startLoop(configJson, tunFd.toLong())
+            ctrl.startLoop(configJson, tunFd)
             controller = ctrl
             Log.i(TAG, "Xray startLoop called, tunFd=$tunFd")
         } catch (e: Exception) {
